Moved LatexCommands/PsiFile extensions from PsiKt to PsiCommandsKt/FilesKt.
