+++ src/nl/rubensten/texifyidea/util/Psi.kt
@@ -1,15 +1,10 @@
 package nl.rubensten.texifyidea.util
 
-import com.intellij.openapi.editor.Document
-import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.psi.PsiComment
-import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.impl.source.tree.LeafPsiElement
 import com.intellij.psi.util.PsiTreeUtil
-import nl.rubensten.texifyidea.index.LatexCommandsIndex
-import nl.rubensten.texifyidea.index.LatexDefinitionIndex
 import nl.rubensten.texifyidea.index.LatexIncludesIndex
 import nl.rubensten.texifyidea.lang.DefaultEnvironment
 import nl.rubensten.texifyidea.lang.Environment
@@ -272,191 +267,8 @@ fun PsiElement.inDirectEnvironmentContext(context: Environment.Context): Boolean
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-//// PSI FILE //////////////////////////////////////////////////////////////////////////////////////////////////////////
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Get the corresponding document of the PsiFile.
- */
-fun PsiFile.document(): Document? = PsiDocumentManager.getInstance(project).getDocument(this)
-
-/**
- * @see [LatexCommandsIndex.getIndexedCommands]
- */
-fun PsiFile.commandsInFile(): Collection<LatexCommands> = LatexCommandsIndex.getItems(this)
-
-/**
- * Get the editor of the file if it is currently opened.
- */
-fun PsiFile.openedEditor() = FileEditorManager.getInstance(project).selectedTextEditor
-
-/**
- * Get all the definitions in the file.
- */
-fun PsiFile.definitions(): Collection<LatexCommands> {
-    return LatexDefinitionIndex.getItems(this)
-            .filter { it.isDefinition() }
-}
-
-/**
- * Get all the definitions and redefinitions in the file.
- */
-fun PsiFile.definitionsAndRedefinitions(): Collection<LatexCommands> {
-    return LatexDefinitionIndex.getItems(this)
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 //// LATEX ELEMENTS ////////////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * Checks whether the given LaTeX commands is a (re)definition or not.
- *
- * This is either a command definition or an environment (re)definition.
- *
- * @return `true` if the command is an environment (re)definition or a command (re)definition, `false` when the command is
- *         `null` or otherwise.
- */
-fun LatexCommands?.isDefinitionOrRedefinition() = this != null &&
-        (this.name in Magic.Command.redefinitions || this.name in Magic.Command.redefinitions)
-
-/**
- * Checks whether the given LaTeX commands is a definition or not.
- *
- * This is either a command definition or an environment definition. Does not count redefinitions.
- *
- * @return `true` if the command is an environment definition or a command definition, `false` when the command is
- *         `null` or otherwise.
- */
-fun LatexCommands?.isDefinition() = this != null && this.name in Magic.Command.definitions
-
-/**
- * Checks whether the given LaTeX commands is a command definition or not.
- *
- * @return `true` if the command is a command definition, `false` when the command is `null` or otherwise.
- */
-fun LatexCommands?.isCommandDefinition(): Boolean {
-    return this != null && ("\\newcommand" == name ||
-            "\\let" == name ||
-            "\\def" == name ||
-            "\\DeclareMathOperator" == name ||
-            "\\renewcommand" == name)
-}
-
-/**
- * Checks whether the given LaTeX commands is an environment definition or not.
- *
- * @return `true` if the command is an environment definition, `false` when the command is `null` or otherwise.
- */
-fun LatexCommands?.isEnvironmentDefinition(): Boolean {
-    return this != null && ("\\newenvironment" == name ||
-            "\\renewenvironment" == name)
-}
-
-/**
- * @see TexifyUtil.getForcedFirstRequiredParameterAsCommand
- */
-fun LatexCommands.firstRequiredParamAsCommand(): LatexCommands? = TexifyUtil.getForcedFirstRequiredParameterAsCommand(this)
-
-/**
- * Get the command that gets defined by a definition (`\let` or `\def` command).
- */
-fun LatexCommands.definitionCommand(): LatexCommands? = nextCommand()
-
-/**
- * Checks whether the command has a star or not.
- */
-fun LatexCommands.hasStar() = childrenOfType(LeafPsiElement::class).any {
-    it.elementType == LatexTypes.STAR
-}
-
-/**
- * Looks for the next command relative to the given command.
- *
- * @return The next command in the file, or `null` when there is no such command.
- */
-fun LatexCommands.nextCommand(): LatexCommands? {
-    val content = parentOfType(LatexContent::class) ?: return null
-    val next = content.nextSiblingIgnoreWhitespace() as? LatexContent ?: return null
-    return next.firstChildOfType(LatexCommands::class)
-}
-
-/**
- * Looks for the previous command relative to the given command.
- *
- * @return The previous command in the file, or `null` when there is no such command.
- */
-fun LatexCommands.previousCommand(): LatexCommands? {
-    val content = parentOfType(LatexContent::class) ?: return null
-    val previous = content.previousSiblingIgnoreWhitespace() as? LatexContent ?: return null
-    return previous.firstChildOfType(LatexCommands::class)
-}
-
-/**
- * @see TexifyUtil.getForcedFirstRequiredParameterAsCommand
- */
-fun LatexCommands.forcedFirstRequiredParameterAsCommand(): LatexCommands? = TexifyUtil.getForcedFirstRequiredParameterAsCommand(this)
-
-/**
- * Get the name of the command that is defined by `this` command.
- */
-fun LatexCommands.definedCommandName() = when (name) {
-    "\\DeclareMathOperator", "\\newcommand" -> forcedFirstRequiredParameterAsCommand()?.name
-    else -> definitionCommand()?.name
-}
-
-/**
- * @see TexifyUtil.isCommandKnown
- */
-fun LatexCommands.isKnown(): Boolean = TexifyUtil.isCommandKnown(this)
-
-/**
- * Get the environment name of a begin/end command.
- *
- * @param element
- *              Either a [LatexBeginCommand] or a [LatexEndCommand]
- */
-private fun beginOrEndEnvironmentName(element: PsiElement) = element.firstChildOfType(LatexNormalText::class)?.text
-
-/**
- * Get the `index+1`th required parameter of the command.
- *
- * @throws IllegalArgumentException When the index is negative.
- */
-@Throws(IllegalArgumentException::class)
-fun LatexCommands.requiredParameter(index: Int): String? {
-    require(index >= 0) { "Index must not be negative" }
-
-    val parameters = requiredParameters
-    if (parameters.isEmpty() || index >= parameters.size) {
-        return null
-    }
-
-    return parameters[index]
-}
-
-/**
- * Finds the indentation of the line where the section command starts.
- */
-fun LatexCommands.findIndentation(): String {
-    val file = containingFile
-    val document = file.document() ?: return ""
-    val lineNumber = document.getLineNumber(textOffset)
-    return document.lineIndentation(lineNumber)
-}
-
-/**
- * If the given command is an include command, the contents of the first argument will be read.
- *
- * @return The included filename or `null` when it's not an include command or when there
- * are no required parameters.
- */
-fun LatexCommands.includedFileName(): String? {
-    if (commandToken.text !in Magic.Command.includes) return null
-    val required = requiredParameters
-    if (required.isEmpty()) return null
-    return required.first()
-}
 
 /**
  * @see TexifyUtil.isEntryPoint
@@ -493,6 +305,14 @@ fun LatexBeginCommand.endCommand(): LatexEndCommand? = nextSiblingOfType(LatexEn
  * Get the environment name of the end command.
  */
 fun LatexEndCommand.environmentName(): String? = beginOrEndEnvironmentName(this)
+
+/**
+ * Get the environment name of a begin/end command.
+ *
+ * @param element
+ *              Either a [LatexBeginCommand] or a [LatexEndCommand]
+ */
+private fun beginOrEndEnvironmentName(element: PsiElement) = element.firstChildOfType(LatexNormalText::class)?.text
 
 /**
  * Finds the [LatexBeginCommand] that matches the end command.
